sensor/requirements.txt
ruuvitag_sensor
paho-mqtt